@@ -1,22 +1,22 @@
 {
-  "name": "classify2",
-  "version": "0.0.3",
+  "name": "classify",
+  "version": "0.1.0",
+  "description": "Use OCLC's Classify2 API",
+  "engines": {
+    "node": "14.x"
+  },
   "main": "index.js",
   "scripts": {
     "test": "node_modules/.bin/mocha"
   },
-  "author": "Douglas Gardner <douglas@chippy.ch>",
-  "license": "BSD",
   "dependencies": {
-    "isbn": "~0.1.0",
-    "request": "^2.88.0",
-    "xml2js": "~0.2.8"
+    "isbn": "^0.4.0",
+    "request": "^2.88.2",
+    "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "mocha": "^6.0.2",
     "should": "~1.3.0"
   },
-  "engines": {
-    "node": ">= 0.8"
-  }
+  "author": "Isaac List",
+  "license": "MPL-2.0"
 }
